Refractor Book.js
