+++ src/components/Book.js
@@ -1,37 +1,14 @@
 import PropTypes from 'prop-types';
-import { useDispatch } from 'react-redux';
-import { removeBook } from '../redux/books/booksReducer';
+import BooData from './BookData';
 import BookCompletion from './BookCompletion';
 import BookChapter from './BookChapter';
 import './Book.css';
 
 const Book = (props) => {
-  const dispatch = useDispatch();
-  const {
-    bookId,
-    title,
-    author,
-    category,
-  } = props;
-
-  const handleRemove = () => {
-    dispatch(removeBook({ item_id: bookId }));
-  };
-
+  const { bookData } = props;
   return (
     <section className="book-section">
-      <div className="flex-2">
-        <div>
-          <span className="category">{category}</span>
-          <h2 className="m-0">{title}</h2>
-          <span className="author">{author}</span>
-        </div>
-        <div>
-          <input className="button pl-0 pr-2" type="button" value="Comments" />
-          <input className="button buttom-center" type="button" value="Remove" onClick={handleRemove} />
-          <input className="button pl-2" type="button" value="Edit" />
-        </div>
-      </div>
+      <BooData bookData={bookData} />
       <BookCompletion />
       <BookChapter />
     </section>
@@ -39,17 +16,11 @@ const Book = (props) => {
 };
 
 Book.propTypes = {
-  bookId: PropTypes.string,
-  title: PropTypes.string,
-  author: PropTypes.string,
-  category: PropTypes.string,
+  bookData: PropTypes.shape(),
 };
 
 Book.defaultProps = {
-  bookId: '',
-  title: '',
-  author: '',
-  category: '',
+  bookData: [],
 };
 
 export default Book;
