+++ src/pages/BookPage.js
@@ -19,10 +19,14 @@ const BookPage = () => {
     (book) => (
       <Book
         key={book.item_id}
-        bookId={book.item_id}
-        title={book.title}
-        author="Anonymous"
-        category={book.category}
+        bookData={
+          {
+            item_id: book.item_id,
+            title: book.title,
+            author: 'Anonymous',
+            category: book.category,
+          }
+        }
       />
     ),
   );
